Addressed review commenta
